# lib/api-zod/src/generated/types/brandUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Notification Engine API
 * OpenAPI spec version: 0.2.0
 */
import type { BrandUpdateCategory } from './brandUpdateCategory';
import type { BrandUpdateTheme } from './brandUpdateTheme';

export interface BrandUpdate {
  name?: string;
  category?: BrandUpdateCategory;
  theme?: BrandUpdateTheme;
  primaryColor?: string;
  /** @nullable */
  headerBg?: string | null;
  cardBg?: string;
  bodyBg?: string;
  textColor?: string;
  /** @nullable */
  accentBg?: string | null;
  /** @nullable */
  logoUrl?: string | null;
}