# lib/api-zod/src/generated/types/emailDispatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Notification Engine API
 * OpenAPI spec version: 0.2.0
 */
import type { EmailDispatchStatus } from './emailDispatchStatus';

export interface EmailDispatch {
  id: number;
  templateId: number;
  templateName: string;
  brandKey: string;
  brandName: string;
  recipientEmail: string;
  emailSubject: string;
  /** @nullable */
  telegramUserId?: string | null;
  status: EmailDispatchStatus;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
}