# lib/api-zod/src/generated/types/emailDispatchList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Notification Engine API
 * OpenAPI spec version: 0.2.0
 */
import type { EmailDispatch } from './emailDispatch';

export interface EmailDispatchList {
  dispatches: EmailDispatch[];
  total: number;
}