# lib/api-zod/src/generated/types/emailPreviewResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Notification Engine API
 * OpenAPI spec version: 0.2.0
 */

export interface EmailPreviewResult {
  html: string;
  variables: string[];
}