# lib/api-zod/src/generated/types/emailSendResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Notification Engine API
 * OpenAPI spec version: 0.2.0
 */

export interface EmailSendResult {
  success: boolean;
  dispatchId: number;
  /** @nullable */
  error?: string | null;
}